Update Modules. Add validation of entered values in getAddProductsFromUser().

+++ src/main/java/entity/Products.java
@@ -3,11 +3,11 @@ package entity;
 /* Класс с параметрами "название товара - name" и "цена - price" для добавления товаров в калькулятор */
 public class Products {
     String name;
-    float price;
+    double price;
     int count;
 
     /* Конструктор класса */
-    public Products(String productsName, float productsPrice, int productsCount) {
+    public Products(String productsName, double productsPrice, int productsCount) {
         this.name = productsName;
         this.price = productsPrice;
         this.count = productsCount;
